+++ examples/server/src/resolvers/Mutation.ts
@@ -1,7 +1,7 @@
-import { ResolverMap } from "../generated/schema";
+import { Resolvers } from "../generated/schema";
 import { MyContext } from "../MyContext";
 
-export const mutationResolvers: ResolverMap<MyContext> = {
+export const mutationResolvers: Resolvers<MyContext> = {
   Mutation: {
     youDied() {
       return true;