+++ examples/server/src/resolvers/index.ts
@@ -1,11 +1,11 @@
-import { ResolverMap } from "../generated/schema";
+import { Resolvers } from "../generated/schema";
 import { MyContext } from "../MyContext";
 import { mutationResolvers } from "./Mutation";
 import { queryResolvers } from "./Query";
 import { reviewResolvers } from "./reviewResolvers";
 import { scalarResolvers } from "./scalars";
 
-export const resolvers: ResolverMap<MyContext>[] = [
+export const resolvers: Resolvers<MyContext>[] = [
   queryResolvers,
   mutationResolvers,
   scalarResolvers,