+++ src/index.ts
@@ -1,5 +1,4 @@
 export * from "./codegen";
-export * from "./execute";
 export * from "./http";
 export * from "./introspect";
 export * from "./read";